oppe2/requirements.txt
flask==2.3.3
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
mlflow==2.7.1
gunicorn==21.2.0
requests==2.31.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
matplotlib==3.7.2
seaborn==0.12.2
